Move proto utils into their own package
